Allow to force evaluation points in plot.gbm
This is useful to evaluate the effects of two models on
the exact same data points. An example application is the
evaluation of the confidence around model effects through
bootstraps. For each bootstrap, the data range will be
slightly different and the points at which plot.gbm will
evaluate the effect of the predictor will also be slightly
different, which complicates the computation of a mean+range
of effects afterwards.

The argument is not checked very much and the description in
the help is a little verbose but that's a start.